+++ easy-common/easy-entity/src/main/java/com/rongyi/easy/activitymanage/entity/GoodsCheckRecord.java
@@ -7,7 +7,7 @@ public class GoodsCheckRecord {
 
     private Integer activityGoodsId;
 
-    private Byte status;
+    private Integer status;
 
     private String record;
 
@@ -31,11 +31,11 @@ public class GoodsCheckRecord {
         this.activityGoodsId = activityGoodsId;
     }
 
-    public Byte getStatus() {
+    public Integer getStatus() {
         return status;
     }
 
-    public void setStatus(Byte status) {
+    public void setStatus(Integer status) {
         this.status = status;
     }
